Convert 'default' values to callables.
Change the 'default' value for Wannier90WorkChain and
PwBandStructureWorkChain to callables, by using a lambda.
The reasons for this change are detailed in aiida-core PR #3466.
In short: Using an already-instantiated AiiDA Node as default
can cause issues, for example if the Node is deleted from the DB
but the object still lives on in the Python interpreter. Using
a callable instead avoids this problem, because the Node is
instantiated when the process is started.