Make 'parameters' input of 'pw2wannier90' optional.
When the `pw2wannier90.parameters` are not set, the workchain
excepts because `inputs.parameters` is accessed without a check.

In `aiida-quantumespresso`, the `parameters` are optional, which
is why this error is not caught by the input validation.

This is fixed by catching the exception, and using an empty
dictionary.

Passing an empty `Dict` is now also allowed, by setting an empty
`parameters['inputpp']` dictionary if it does not exist.