Add `remote_symlink_files` to `Wannier90BaseWorkChain.inputs.settings`
`remote_symlink_files` accepts a list of filename which will be
symlinked from the input remote folder to the job submission folder.
This allows the workflow to symlink files generated by an intermediate
pw2wannier90 run to the wannier90 folder, without knowing the remote
folder of pw2wannier90 when creating the builder. Note the
`Wannier90Calculation` has an `additional_remote_symlink_list` in its
`inputs.settings`, however that requires knowing the absolute path of
the remote folder when filling in the inputs of the
`Wannier90Calculation`. This is impossible in the bands workchain since
the remote folder of pw2wannier90 is created on the fly.
Now the inputs of `Wannier90BaseWorkChain` is exposed in the
`Wannier90BandsWorkChain`, instead of the inputs of
`Wannier90Calculation`. This allows specifying the
`remote_symlink_files` of the `Wannier90BaseWorkChain`.

If an `inputs.metadata.options.stash` is provided, in wannier90_pp step,
the stash option is removed, this avoids creating an unecessary
`RemoteStashFolderData` in wannier90_pp. In wannier90 step, the stash
option is restored, thus still allowing stashing files.