Add `Wannier90OptimizeWorkChain`
The workchain will iterate a list of `dis_proj_min/max` and compute
bands distance on-the-fly, if the bands distance is under the specified
threshold, the workchain stops the iteration and computes bands and
plots wannier functions on top of the optimal wannier functions.

A CLI tool is added to show the checkerboard plot of combinations of
`dis_proj_min/max` and help identifing the optimal parameters.