✨ Add script to generate PseudoDojo semicore JSON files
Add a script to automatically generate the JSON files that indicate the semicore
orbitals. It uses some very rudimentary heuristics to determine which of the
pseudo wave functions should be considered semicore, tuned to properly
regenerate the existing JSON files for PseudoDojo v0.4 LDA and PBE "standard".

The script is added to a new `dev` directory, which can be used to house scripts
that are only really useful to developers/maintainers of the package. In line
with this, a `dev` optional dependency list is set up in the `pyproject.toml`
for packages required to run these development tools.